requirements.txt
appdirs==1.4.3
arrow==0.10.0
beautifulsoup4==4.6.0
disqus==0.0.4
django
django-disqus==0.5
django-emoticons==1.2
django-forms-bootstrap==3.1.0
django-bootstrap-form==3.4
django-registration==2.2
django-bootstrap-datepicker-widget==0.1
django-contact-form==1.6
django-tinymce
djangorestframework==3.6.3
funcsigs==1.0.2
gunicorn==19.9.0
mock==2.0.0
pbr==3.1.1
Pillow==4.1.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.20.0
six==1.10.0
django-simple-captcha
django_crispy_forms==1.7.2
mysql-python
django-debug-toolbar